Restructure Intro component directory
Moved Intro component and related files to a new "_intro" directory for better organization. This change involved updating import paths and renaming associated stylesheet files, ensuring the application structure is cleaner and more maintainable.
