+++ frontend-next-migration/src/app/[lng]/(home)/_intro/Intro.module.scss
@@ -4,7 +4,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background: url('') no-repeat center center;
+  //background: url('') no-repeat center center;
   background-size: cover;
   z-index: 1;
   transition: opacity 0.5s ease, visibility 0.5s ease;
