+++ frontend-next-migration/src/app/[lng]/(home)/_intro/Intro.tsx
